+++ src/types.ts
@@ -1,5 +1,6 @@
 import 'ultimate-crosscode-typedefs'
 import type { Mod } from 'ultimate-crosscode-typedefs/modloader/mod'
+import 'cc-instanceinator/src/plugin'
 
 export type Mod1 = Writable<Mod> & {
     isCCModPacked: boolean