+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 # Change Log
 
 ## [Unreleased]
+
+### Changed
+
+- Reset ig.AnimatedEntity animation timers on screeenshot
+
+### Fixed
+
+- Fix compatibility with cc-multibakery
+
 ## [1.0.0] 2026-08-18
 
 Initial release